+++ core-application/src/main/java/gohigher/user/service/UserCommandService.java
@@ -1,4 +1,4 @@
-package gohigher.user.usecase;
+package gohigher.user.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
